back to using covariate names for columns, and tests pass, and smoke test on meid 2005
