+++ examples/simple_load_from_epiviz.py
@@ -8,7 +8,9 @@ from cascade.core.db import latest_model_version
 from cascade.input_data.db.configuration import settings_json_from_epiviz
 from cascade.executor.no_covariate_main import bundle_to_observations
 from cascade.input_data.configuration.form import Configuration
-from cascade.input_data.db.bundle import bundle_with_study_covariates, freeze_bundle
+from cascade.input_data.configuration.construct_bundle import \
+    bundle_with_study_covariates
+from cascade.input_data.db.bundle import freeze_bundle
 from cascade.dismod.serialize import model_to_dismod_file
 from cascade.input_data.configuration.builder import (
     initial_context_from_epiviz,
