+++ src/cascade/executor/epiviz_runner.py
@@ -14,9 +14,11 @@ from cascade.input_data.db.configuration import load_settings
 from cascade.input_data.db.csmr import load_csmr_to_t3
 from cascade.input_data.db.asdr import load_asdr_to_t3
 from cascade.input_data.db.mortality import get_cause_specific_mortality_data, get_age_standardized_death_rate_data
-from cascade.executor.no_covariate_main import bundle_to_observations, build_constraint
 from cascade.executor.dismod_runner import run_and_watch, DismodATException
-from cascade.input_data.db.bundle import bundle_with_study_covariates, freeze_bundle
+from cascade.input_data.configuration.construct_bundle import \
+    bundle_with_study_covariates, bundle_to_observations
+from cascade.input_data.configuration.construct_study import add_special_study_covariates
+from cascade.input_data.db.bundle import freeze_bundle
 from cascade.dismod.serialize import model_to_dismod_file
 from cascade.model.integrands import make_average_integrand_cases_from_gbd
 from cascade.saver.save_model_results import save_model_results
@@ -25,9 +27,9 @@ from cascade.input_data.configuration.builder import (
     initial_context_from_epiviz,
     fixed_effects_from_epiviz,
     random_effects_from_epiviz,
-)
+    build_constraint)
 
-from cascade.core.log import getLoggers
+from cascade.core import getLoggers
 CODELOG, MATHLOG = getLoggers(__name__)
 
 
@@ -144,7 +146,6 @@ def model_context_from_settings(execution_context, settings):
     )
     bundle = bundle.query("location_id == @execution_context.parameters.location_id")
     observations = bundle_to_observations(model_context.parameters, bundle)
-    observations = observations.rename(columns={"location_id": "node_id"})
     model_context.input_data.observations = observations
 
     mask = model_context.input_data.observations.standard_error > 0
@@ -157,6 +158,7 @@ def model_context_from_settings(execution_context, settings):
     add_mortality_data(model_context, execution_context)
     add_omega_constraint(model_context, execution_context)
     model_context.average_integrand_cases = make_average_integrand_cases_from_gbd(execution_context)
+    add_special_study_covariates(study_covariate_records, bundle, model_context.average_integrand_cases)
 
     fixed_effects_from_epiviz(model_context, study_covariate_records, execution_context, settings)
     random_effects_from_epiviz(model_context, settings)
